+++ packages/orq-rc/src/models/components/evaluatorresponsejsonschema.ts
@@ -32,8 +32,8 @@ export const EvaluatorResponseJsonSchema$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-08-26T09:46:51.830Z"),
-  updated: z.string().default("2026-08-26T09:46:51.831Z"),
+  created: z.string().default("2026-08-26T11:10:54.428Z"),
+  updated: z.string().default("2026-08-26T11:10:54.428Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   project_id: z.string().optional(),
   guardrail_config: z.any().optional(),