+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-26T09:46:50.254Z",
+    "2026-08-26T11:10:52.849Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {